refactor: GenreRecommender를 Domain Layer로 분리
